LPS-50031 Adapt to new ways
