+++ modules/util/osgi-util/build.gradle
@@ -9,7 +9,7 @@ dependencies {
 	testIntegrationCompile group: "org.apache.felix", name: "org.apache.felix.framework", version: "4.4.0"
 	testIntegrationCompile group: "org.apache.felix", name: "org.apache.felix.main", version: "4.4.0"
 	testIntegrationCompile group: "org.jboss.arquillian.container", name: "arquillian-container-felix-embedded", version: "2.1.0.CR15"
-	testIntegrationCompile group: "org.jboss.arquillian.junit", name: "arquillian-junit-container", version: "1.1.6.Final"
+	testIntegrationCompile group: "org.jboss.arquillian.junit", name: "arquillian-junit-container", version: "1.1.3.Final"
 }
 
 liferay {
