+++ modules/util/osgi-util/src/testIntegration/java/com/liferay/osgi/util/test/ReflectionServiceTrackerTestUtil.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.osgi.util.test;
+
+import java.util.Dictionary;
+import java.util.Hashtable;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceRegistration;
+
+/**
+ * @author Carlos Sierra Andrés
+ */
+public class ReflectionServiceTrackerTestUtil {
+
+	protected static <T> ServiceRegistration<T> registerServiceWithRanking(
+		BundleContext bundleContext, Class<T> clazz, T service, int ranking) {
+
+		Dictionary<String, Integer> properties =
+			new Hashtable<String, Integer>();
+
+		properties.put("service.ranking", ranking);
+
+		return bundleContext.registerService(clazz, service, properties);
+	}
+
+}
